application/requirements.txt
Flask==3.1.0
numpy==2.1.3
opencv_python_headless==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
segment_anything==1.0
torch==2.4.0
ultralytics==8.3.25